Bram's changes to line_tracker.cpp
For most of the changes made I have left your (probably correct) code in and just commented it out. 

The only place where I haven't done this is for the number of inputs for the line sensors, as I am fairly sure we only need one pin assigned to each as they as have one output wire.

I have also written my own version line follower code just so we have a couple of options to test when we have access to the robot again. I'm yet to work out how to handle intersections so that could be something we look at working through together. Please message me if you have any questions